requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
greenlet==3.1.0
redis==5.1.1
celery[redis]==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
openpyxl==3.1.5
aiohttp==3.10.8
python-dateutil==2.9.0